Move AsFloat and AnchorPoint into arcade.types submodules
