+++ arcade/types/rect.py
@@ -3,9 +3,13 @@
 from __future__ import annotations
 
 from typing import NamedTuple, Optional, TypedDict
+
+
 from pyglet.math import Vec2
 
-from arcade.types import AsFloat
+from arcade.types.numbers import AsFloat
+from arcade.types.vector_like import AnchorPoint
+
 from arcade.utils import ReplacementWarning, warning
 
 RectParams = tuple[AsFloat, AsFloat, AsFloat, AsFloat]
@@ -21,19 +25,6 @@ class RectKwargs(TypedDict):
     height: float
     x: float
     y: float
-
-
-class AnchorPoint:
-    """Provides helper aliases for several Vec2s to be used as anchor points in UV space."""
-    BOTTOM_LEFT = Vec2(0.0, 0.0)
-    BOTTOM_CENTER = Vec2(0.5, 0.0)
-    BOTTOM_RIGHT = Vec2(1.0, 0.0)
-    CENTER_LEFT = Vec2(0.0, 0.5)
-    CENTER = Vec2(0.5, 0.5)
-    CENTER_RIGHT = Vec2(1.0, 0.5)
-    TOP_LEFT = Vec2(0.0, 1.0)
-    TOP_CENTER = Vec2(0.5, 1.0)
-    TOP_RIGHT = Vec2(1.0, 1.0)
 
 
 class Rect(NamedTuple):
